+++ sdk/js/packages/node/test/options.test.ts
@@ -117,17 +117,18 @@ describe('when api url is given', () => {
     expect(existing.api_url).toEqual('http://127.0.0.1/api');
   });
 
-  test('reject conflicting input', () => {
+  test('accepting conflicting input', () => {
     const existing: NodeOptions = {};
     const incoming: NodeOptions = {
       api_key: 'some-key',
       api_url: 'http://localhost:8080/@/team/suite/version',
       suite: 'some-other-version'
     };
-    expect(() => update_options(existing, incoming)).toThrowError(
-      'option "suite" is in conflict with provided api_url'
-    );
+    expect(() => update_options(existing, incoming)).not.toThrow();
     expect(existing.api_url).toEqual('http://localhost:8080/');
+    expect(existing.team).toEqual('team');
+    expect(existing.suite).toEqual('suite');
+    expect(existing.version).toEqual('version');
   });
 });